Extract `update_forward_refs_recursive` to module.
