+++ phanas_pydantic_helpers/helpers/__init__.py
@@ -1,3 +1,4 @@
 from .factory import *
 from .helpers import *
+from .misc import *
 from .validators import *
